pyproject.toml
[tool.poetry]
name = "emart23"
version = "0.1.0"
description = "A sophisticated multi-vendor e-commerce web application"
authors = ["NumanIbnMazid <numanibnmazid@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.8"
Django = "^3.2.9"
dj-database-url = "^0.5.0"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
django-safedelete = "^1.1.0"
psycopg2-binary = "^2.9.1"
python-dotenv = "^0.19.1"
python-memcached = "^1.59"
redis = "^3.5.3"
uvicorn = "^0.15.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
